+++ src/sn/CmnLib.ts
@@ -98,6 +98,7 @@ export interface IEvtMng {
 	popLocalEvts(): IHEvt2Fnc;
 	pushLocalEvts(a: IHEvt2Fnc): void;
 	waitLimitedEvent(hArg: HArg, fnc: ()=> void): boolean;
+	hideHint(): void;
 	cvsResize(): void;
 	noticeCompTxt(): void;